pyrit_tests/requirements.txt
# PyRIT requirements (0.10.0)
pyrit>=0.10.0
python-dotenv>=1.0.0
azure-ai-inference>=1.0.0
azure-identity>=1.15.0
openai>=1.0.0

# Optional: For Azure Speech Service (Demo 4)
azure-cognitiveservices-speech>=1.35.0

# Optional: For enhanced visualization
pandas>=2.0.0
matplotlib>=3.7.0
